Tweak chromedp fetch test
When run with the "test" URL, it loads http://127.0.0.1:1234/campfire-commerce/
and actually verify's the content (like the existing dump tests). runner/main.go
uses the "test" url. This make the test a bit more accurate - making sure we
got the expected content instead of just any content. It also helps to reduce
the output noise by not dumping the html to stdout.