Changes the form runner to use a page-per-test
Previously, we were reusing the same page for all cases. But that causes an
issue with https://github.com/lightpanda-io/browser/pull/1130 and, from Google
AI answer, seems to be a common cause of various errors (network, timeouts, ...)

For us, the specific problem is that puppeteer registers an
IntersectionObserver (why?), but the scrip then does another page.goto,
preventing the IntersectionObserver from ever calling back. This ends up
blocking puppeteer.

It wasn't an issue before, because we fired the callback synchronously. But that
was not a correct behavior. This is probably something we should support, but
it isn't clear how. Does Chrome keep the "old" page context around to fulfill
these events?